Move more tests out
